requirements.txt
python==3.7.11
tqdm==4.62.3
numpy==1.21.4  
pytorch==1.10.2
tensorboardX==2.4
torchvision==0.11.3
h5py==3.5.0